refactor: apply simplification review across the branch diff
