+++ crates/ffi/Cargo.toml
@@ -32,6 +32,7 @@ agent-desktop-linux = { path = "../linux" }
 [target.'cfg(target_os = "windows")'.dev-dependencies]
 windows-sys = { version = "0.61", features = [
   "Win32_Foundation",
+  "Win32_System_ApplicationInstallationAndServicing",
   "Win32_System_LibraryLoader",
   "Win32_System_Threading",
   "Win32_UI_WindowsAndMessaging",
